factor: stop truncating the sieve interval index to 32 bits
interval_lo is a uint64 (factor.c:691, computed at :1882 as kmin/((uint64)len << TF_CLSHIFT)),
but get_startval() declared the parameter uint32, so the value was silently truncated at both
call sites. Inside, it scales the small-prime sieve start offsets:

	dstartval = (interval_lo*dstartval) % curr_p;

A truncated interval_lo yields the wrong start offset for every small prime, so the wrong
bit_map bits get cleared and candidates that should be tested are sieved away. That is a missed
factor with no diagnostic - the run reports "no factor found" and exits 0.

The truncation bites once kmin > 2^32 * (len << TF_CLSHIFT): about 7.0e16 for the mod-60 class
set and 4.5e18 for mod-4620. Both are inside the deep 96-bit range Mfactor supports.

Widening the parameter alone would be wrong. The comment above the multiply already says what is
needed:

	where we'll probably want to do a mod-curr_p of interval_lo prior to
	the multiply by (k*curr_p - bit_len) to keep intermediates < (curr_p)^2

but the reduction was never actually done - it was unnecessary only because the truncation kept
interval_lo below 2^32. With a full-width interval_lo the product overflows uint64 and gives a
different wrong answer. So this widens the parameter and reduces mod curr_p before the multiply,
as specified. dstartval <= curr_p and interval_lo%curr_p < curr_p, both under 2^32, so the
product stays below 2^64.

Checked arithmetically: for curr_p = 4294967291 (largest prime below 2^32) the old and new forms
agree for every interval_lo < 2^32 and diverge from 2^32 up, and the naive widening without the
mod-first reduction diverges from both. A stock Mfactor run is unchanged line for line apart
from timings.

+++ src/factor.h
@@ -506,7 +506,7 @@ void	get_startval(
 	const uint64*two_p,	// Here, need the full multiword array (of which use just LSW if P!WORD def'd)
 	const uint32 lenP,	// Manyword case
 	const uint32 bit_len,
-	const uint32 interval_lo, const uint32 incr,
+	const uint64 interval_lo, const uint32 incr,
 	const uint32 nclear, const uint32 nprime, const uint32 p_last_small,
 	const uint8 *pdiff,
 	      uint32*startval
